Split metadynamics potential module
