+++ openxps/__init__.py
@@ -3,14 +3,14 @@ Extended Phase-Space Methods with OpenMM
 """
 
 from ._version import __version__  # noqa: F401
-from .biasing import MetadynamicsPotential  # noqa: F401
 from .context import ExtendedSpaceContext  # noqa: F401
 from .extension_writer import ExtensionWriter  # noqa: F401
 from .extra_dof import ExtraDOF  # noqa: F401
+from .metadynamics_potential import MetadynamicsPotential  # noqa: F401
 
 __all__ = [
-    "MetadynamicsPotential",
     "ExtendedSpaceContext",
     "ExtensionWriter",
     "ExtraDOF",
+    "MetadynamicsPotential",
 ]
